Complete OoeyTerminal interface
The OoeyTerminal works correctly at printing new output and new input
and handles all immediately obvious edge cases. However, the internal
implementation can only handle having one thing in the input or output
not handled at a time. This is because it is implemented with MVars,
which only allow 1 value at a time, not a propre queue. We'll need to
change that to a TChan from the stm package most likely.